Tilgangskontroll på sanity preview routes
Kun betabrukere og team mulighetsrommet

mao admin-flate access
