+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/plugins/Authentication.kt
@@ -18,6 +18,7 @@ enum class AuthProvider {
     AZURE_AD_TEAM_MULIGHETSROMMET,
     AZURE_AD_DEFAULT_APP,
     AZURE_AD_TILTAKSGJENNOMFORING_APP,
+    AZURE_AD_ADMIN_FLATE_TILGANG,
 }
 
 object AppRoles {
@@ -67,6 +68,27 @@ fun Application.configureAuthentication(
                 }
 
                 if (!hasNavAnsattRoles(credentials, NavAnsattRolle.TEAM_MULIGHETSROMMET)) {
+                    return@validate null
+                }
+
+                JWTPrincipal(credentials.payload)
+            }
+        }
+
+        jwt(AuthProvider.AZURE_AD_ADMIN_FLATE_TILGANG.name) {
+            verifier(jwkProvider, azure.issuer) {
+                withAudience(azure.audience)
+            }
+
+            validate { credentials ->
+                credentials["NAVident"] ?: run {
+                    application.log.warn("Access denied. Mangler claim 'NAVident'.")
+                    return@validate null
+                }
+
+                if (!hasNavAnsattRoles(credentials, NavAnsattRolle.TEAM_MULIGHETSROMMET) &&
+                    !hasNavAnsattRoles(credentials, NavAnsattRolle.BETABRUKER)
+                ) {
                     return@validate null
                 }
