+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/routes/v1/SanityPreviewRoutes.kt
@@ -1,0 +1,44 @@
+package no.nav.mulighetsrommet.api.routes.v1
+
+import io.ktor.server.application.*
+import io.ktor.server.auth.*
+import io.ktor.server.request.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+import io.ktor.server.util.*
+import kotlinx.serialization.Serializable
+import no.nav.mulighetsrommet.api.plugins.getNavAnsattAzureId
+import no.nav.mulighetsrommet.api.services.PoaoTilgangService
+import no.nav.mulighetsrommet.api.services.VeilederflateService
+import org.koin.ktor.ext.inject
+
+fun Route.sanityPreviewRoutes() {
+    val veilederflateService: VeilederflateService by inject()
+    val poaoTilgangService: PoaoTilgangService by inject()
+
+    route("/api/v1/internal/sanity") {
+        post("/tiltaksgjennomforinger/preview") {
+            val request = call.receive<GetRelevanteTiltaksgjennomforingerPreviewRequest>()
+            poaoTilgangService.verfiyAccessToModia(getNavAnsattAzureId())
+
+            val result = veilederflateService.hentPreviewTiltaksgjennomforinger(request)
+            call.respond(result)
+        }
+        get("/tiltaksgjennomforing/preview/{id}") {
+            poaoTilgangService.verfiyAccessToModia(getNavAnsattAzureId())
+            val id = call.parameters.getOrFail("id")
+            val result = veilederflateService.hentPreviewTiltaksgjennomforing(
+                id,
+            )
+            call.respond(result)
+        }
+    }
+}
+
+@Serializable
+data class GetRelevanteTiltaksgjennomforingerPreviewRequest(
+    val innsatsgruppe: String? = null,
+    val tiltakstypeIds: List<String>? = null,
+    val search: String? = null,
+    val geografiskEnhet: String? = null,
+)
